// src/api/types.ts
import { UserAddress } from './mdp/types';

export type CmsTokens = Nullable<{
  isCalculationSuccessful: boolean;
  normalRetirementAge: number;
  earliestRetirementAge: number;
  earliestRetirementDate: string;
  latestRetirementAge: number;
  selectedRetirementAge: number;
  selectedRetirementDate: string;
  chosenLtaPercentage: number;
  remainingLtaPercentage: number;
  ltaLimit: number;
  email: string;
  phoneNumber: string;
  retirementJourneyExpirationDate: string;
  gmpAgeYears: number;
  gmpAgeMonths: number;
  pre88GMPAtGMPAge: number;
  post88GMPAtGMPAge: number;
  post88GMPIncreaseCap: number;
  statePensionDeduction: number;
  pa04Date: string;
  systemDate: string;
  normalMinimumPensionAgeMonths: number;
  normalMinimumPensionAgeYears: number;
  insuranceNumber: string;
  dateOfBirth: string;
  address: UserAddress;
  selectedOptionNumber: number;
  transferReplyByDate: string;
  transferGuaranteeExpiryDate: string;
  transferGuaranteePeriodMonths: number;
  transferQuoteRunDate: string;
  earliestStartRaDateForSelectedDate: string;
  latestStartRaDateForSelectedDate: string;
  timeToNormalRetirementIso: string;
  pensionPaymentDay: number;
  bereavementCaseNumber: string;
  name: string;
  normalRetirementDate: string;
  targetRetirementDate: string;
  targetRetirementAgeIso: string;
  timeToTargetRetirementIso: string;
  currentAgeIso: string;
  totalAVCFund: number;
  totalPension: number;
  dcBalance: number;
  dcProjectedBalance: number;
  ageAtNormalRetirementIso: string;
  submissionDate: string;
}>;